backend/requirements.txt
-e git+https://github.com/EchoVoice-AI/EchoVoice-AI.git@2ff0886acd8980c7f1bb5086c90ce8165ff7239c#egg=agent&subdirectory=backend
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
azure-core==1.36.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
blockbuster==1.5.25
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
cryptography==44.0.3
distro==1.9.0
fastapi==0.121.3
forbiddenfruit==0.1.4
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
isodate==0.7.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain-core==1.1.0
langchain-openai==1.1.0
langgraph==1.0.3
langgraph-api==0.5.24
langgraph-checkpoint==3.0.1
langgraph-cli==0.4.7
langgraph-prebuilt==1.0.5
langgraph-runtime-inmem==0.18.1
langgraph-sdk==0.2.9
langsmith==0.4.46
Mako==1.3.10
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
openai==2.8.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pluggy==1.6.0
protobuf==6.33.1
psycopg==3.2.13
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
ruff==0.14.6
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
sse-starlette==2.1.3
starlette==0.50.0
structlog==25.5.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
truststore==0.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
wheel==0.45.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
